requirements.txt
flask
google-genai
python-dotenv
youtube-transcript-api
